Add --dry-run flag to not start Flow

+++ cfde_client/cfde_client/main.py
@@ -22,11 +22,12 @@ def cli():
 @click.option("--output-dir", default=None, show_default=True, type=click.Path(exists=False))
 @click.option("--delete-dir/--keep-dir", is_flag=True, default=False, show_default=True)
 @click.option("--ignore-git/--handle-git", is_flag=True, default=False, show_default=True)
+@click.option("--dry-run", is_flag=True, default=False, show_default=True)
 # TODO: Debug "hidden" missing parameter
 @click.option("--server", default=None)  # , hidden=True)
 @click.option("--bag-kwargs-file", type=click.Path(exists=True), default=None)  # , hidden=True)
 @click.option("--client-state-file", type=click.Path(exists=True), default=None)  # , hidden=True)
-def run(data_path, catalog, schema, output_dir, delete_dir, ignore_git,
+def run(data_path, catalog, schema, output_dir, delete_dir, ignore_git, dry_run,
         server, bag_kwargs_file, client_state_file):
     """Start the Globus Automate Flow to ingest CFDE data into DERIVA."""
     if bag_kwargs_file:
@@ -42,7 +43,7 @@ def run(data_path, catalog, schema, output_dir, delete_dir, ignore_git,
         start_res = cfde.start_deriva_flow(data_path, catalog_id=catalog, schema=schema,
                                            output_dir=output_dir, delete_dir=delete_dir,
                                            handle_git_repos=(not ignore_git),
-                                           server=server, **bag_kwargs)
+                                           server=server, dry_run=dry_run, **bag_kwargs)
     except Exception as e:
         print("Error while starting Flow: {}".format(str(e)))
         return
@@ -50,8 +51,9 @@ def run(data_path, catalog, schema, output_dir, delete_dir, ignore_git,
         if not start_res["success"]:
             print("Error during Flow startup: {}".format(start_res["error"]))
         else:
-            with open(client_state_file, 'w') as out:
-                json.dump(start_res, out)
+            if not dry_run:
+                with open(client_state_file, 'w') as out:
+                    json.dump(start_res, out)
             print(start_res["message"])
